requirements.txt
attrs==19.3.0
more-itertools==8.2.0
packaging==20.3
pluggy==0.13.1
py==1.8.1
pyparsing==2.4.6
pytest==5.4.1
six==1.14.0
wcwidth==0.1.8
grpcio
grpcio-tools
PyYAML==5.3.1
django


# dev reqs
ipython
twine


# to reload deps
# pip install pytest
